//interconversion of two HashMap
In this code, the java.util.HashMap class is imported, which provides the implementation of a HashMap data structure.

In the main method, a new HashMap hm1 is created with key-value pairs of type Integer and String. Three key-value pairs are added to hm1.

Then, a new HashMap hm2 is created and initialized with the contents of hm1 using the constructor HashMap<>(hm1). This constructor creates a new HashMap with the same mappings as the specified HashMap.

Finally, the mappings of both hm1 and hm2 are printed using System.out.println() statements.